split typechecker into its own file
